Move classes/modules into gem namespace and give them standard library-like aliases
